tests/wpt/web-platform-tests/css/css-sizing/parsing/min-height-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Intrinsic & Extrinsic Sizing Test: parsing min-height with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-min-height">
<meta name="assert" content="min-height supports only the grammar 'auto | <length-percentage> | min-content | max-content | fit-content(<length-percentage>)'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("min-height", "none");
test_invalid_value("min-height", "complex");

test_invalid_value("min-height", "-10%");
test_invalid_value("min-height", "-0.5em");
test_invalid_value("min-height", "60");

test_invalid_value("min-height", "10px 20%");
test_invalid_value("min-height", "max-content 10px");
test_invalid_value("min-height", "min-content max-content");

// The following were previously supported by a min-height definition in
// https://drafts.csswg.org/css-box-3/
test_invalid_value("min-height", "available");
test_invalid_value("min-height", "10px border-box");
test_invalid_value("min-height", "content-box 20%");
test_invalid_value("min-height", "fit-content");
</script>
</body>
</html>
